Add wait option to allow delaying between requests
This change introduces a new `wait` field to the EndpointConfig struct
that allows configuring a delay between API requests. This can be
useful for rate limiting or to avoid overwhelming the API endpoint
with too many rapid requests.

The wait time is specified in milliseconds and is applied after each
successful API request before returning the response.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>