Add Gemini 2.5 Flash synthmerge_bench results
This is the first model tested where the added context makes it
underpform. It's the fastest model around 779 t/s with context
enabled.

Unlike 2.5 Pro and 3 Pro preview, "reasoning_effort: low" doesn't seem
to work well with 2.5 Flash.

The accuracy aligned and stripped is better than Gemini 2.5 Pro, but
the Gemini 2.5 Pro score in the README file was achieved on a older
codebase that lacked various improvements.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>